[abi] move emulator to a separate package
